Transizioni permesse solo per sprint attivi Aggiunta possibilità di modificare item nel backlog Corretto bug modifica scrum product workflow

+++ src/main/java/com/isssr/ticketing_system/controller/ScrumProductWorkflowController.java
@@ -1,10 +1,14 @@
 package com.isssr.ticketing_system.controller;
 
 import com.isssr.ticketing_system.dao.ScrumProductWorkflowDao;
+import com.isssr.ticketing_system.dao.TargetDao;
 import com.isssr.ticketing_system.dto.ScrumProductWorkflowDto;
 import com.isssr.ticketing_system.entity.ScrumProductWorkflow;
+import com.isssr.ticketing_system.entity.Sprint;
+import com.isssr.ticketing_system.entity.Target;
 import com.isssr.ticketing_system.exception.NotFoundEntityException;
 import com.isssr.ticketing_system.exception.ScrumProductWorkflowNotSavedException;
+import com.isssr.ticketing_system.exception.UpdateException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,6 +22,8 @@ public class ScrumProductWorkflowController {
 
     @Autowired
     ScrumProductWorkflowDao scrumProductWorkflowDao;
+    @Autowired
+    TargetDao targetDao;
 
     public ScrumProductWorkflowDto addScrumProductWorkflow(ScrumProductWorkflowDto workflowDto)
             throws ScrumProductWorkflowNotSavedException {
@@ -44,23 +50,45 @@ public class ScrumProductWorkflowController {
         return scrumProductWorkflowDtos;
     }
 
-    public void removeScrumProductWorkflow(Long scrumProductWorkflowId) throws NotFoundEntityException {
+    public void removeScrumProductWorkflow(Long scrumProductWorkflowId) throws NotFoundEntityException, UpdateException {
         Optional<ScrumProductWorkflow> scrumProductWorkflow = scrumProductWorkflowDao.findById(scrumProductWorkflowId);
         if (!scrumProductWorkflow.isPresent()){
             throw new NotFoundEntityException();
         }
+
+        // Se esiste uno prodotto con il workflow selezionato l'eliminazione viene impedita
+        List<Target> targetWithSelectedWorkflow =  targetDao.findAllByScrumProductWorkflow(scrumProductWorkflow.get());
+        if (!targetWithSelectedWorkflow.isEmpty()){
+            throw new UpdateException();
+        }
+
         scrumProductWorkflowDao.delete(scrumProductWorkflow.get());
     }
 
     public ScrumProductWorkflowDto updateScrumProductWorkflow(ScrumProductWorkflowDto scrumProductWorkflowDto)
-            throws NotFoundEntityException {
+            throws NotFoundEntityException, UpdateException {
+
         Optional<ScrumProductWorkflow> scrumProductWorkflow = scrumProductWorkflowDao.findById(scrumProductWorkflowDto.getId());
         if (!scrumProductWorkflow.isPresent()){
             throw new NotFoundEntityException();
         }
+
+        // Se esiste uno prodotto con il workflow selezionato avente uno sprint attivo, l'operazione di modifica viene
+        // impedita e un opportuna eccezione viene sollevata
+        List<Target> targetWithSelectedWorkflow =  targetDao.findAllByScrumProductWorkflow(scrumProductWorkflow.get());
+        for (Target target : targetWithSelectedWorkflow){
+            for(Sprint sprint : target.getSprints()){
+                if (sprint.getIsActive() != null && sprint.getIsActive()){
+                    throw new UpdateException();
+                }
+            }
+        }
+
+        // Altrimenti viene effettuata la modifica
         scrumProductWorkflow.get().setName(scrumProductWorkflowDto.getName());
         scrumProductWorkflow.get().setStates(scrumProductWorkflowDto.getStates());
         scrumProductWorkflowDao.save(scrumProductWorkflow.get());
         return scrumProductWorkflowDto;
     }
+
 }
